Check before detaching tags if deleted entity was soft deleted

+++ src/Traits/Taggable.php
@@ -104,7 +104,8 @@ trait Taggable
     public static function bootTaggable()
     {
         static::deleted(function (self $model) {
-            $model->tags()->detach();
+            // Check if this is a soft delete or not by checking if `SoftDeletes::isForceDeleting` method exists
+            (method_exists($model, 'isForceDeleting') && ! $model->isForceDeleting()) || $model->tags()->detach();
         });
     }
